update logic ใหม่ ตาม functional
