--- src/index/hooks/Calendar.ts
@@ -1,3 +1,0 @@
-export const sayHello = () => {
-    console.log('sayHello :>> ')
-}
